snRuntime: Faster HW barrier implementation
